+++ storm-core/test/clj/org/apache/storm/drpc_test.clj
@@ -22,11 +22,14 @@
   (:import [org.apache.storm.coordination CoordinatedBolt$FinishedCallback])
   (:import [org.apache.storm LocalDRPC LocalCluster])
   (:import [org.apache.storm.tuple Fields])
+  (:import [org.mockito Mockito])
+  (:import [org.mockito.exceptions.base MockitoAssertionError])
   (:import [org.apache.storm.utils ConfigUtils]
            [org.apache.storm.utils.staticmocking ConfigUtilsInstaller])
   (:import [org.apache.storm.generated DRPCExecutionException])
   (:import [java.util.concurrent ConcurrentLinkedQueue])
   (:import [org.apache.storm Thrift])
+  (:import [org.apache.storm.daemon DrpcServer])
   (:use [org.apache.storm config testing])
   (:use [org.apache.storm.internal clojure])
   (:use [org.apache.storm.daemon common drpc])
@@ -231,24 +234,24 @@
         delay-seconds 2
         conf {DRPC-REQUEST-TIMEOUT-SECS delay-seconds}
         mock-cu (proxy [ConfigUtils] []
-                  (readStormConfigImpl [] conf))]
+                  (readStormConfigImpl [] conf))
+        drpc-handler (proxy [DrpcServer] [conf]
+                       (acquireQueue [function] queue))]
     (with-open [_ (ConfigUtilsInstaller. mock-cu)]
-      (stubbing [acquire-queue queue]
-        (let [drpc-handler (service-handler conf)]
-          (is (thrown? DRPCExecutionException
+      (is (thrown? DRPCExecutionException
             (.execute drpc-handler "ArbitraryDRPCFunctionName" "")))
-          (is (= 0 (.size queue))))))))
+      (is (= 0 (.size queue))))))
 
-(deftest test-drpc-timeout-cleanup 
+(deftest test-drpc-timeout-cleanup
   (let [queue (ConcurrentLinkedQueue.)
         delay-seconds 1
         conf {DRPC-REQUEST-TIMEOUT-SECS delay-seconds}
         mock-cu (proxy [ConfigUtils] []
-                  (readStormConfigImpl [] conf))]
+                  (readStormConfigImpl [] conf))
+        drpc-handler (proxy [DrpcServer] [conf]
+          (acquireQueue [function] queue)
+          (getTimeoutCheckSecs [] delay-seconds))]
     (with-open [_ (ConfigUtilsInstaller. mock-cu)]
-          (stubbing [acquire-queue queue
-                     timeout-check-secs delay-seconds]
-              (let [drpc-handler (service-handler conf)]
-                (is (thrown? DRPCExecutionException 
-                             (.execute drpc-handler "ArbitraryDRPCFunctionName" "no-args"))))))))
+      (is (thrown? DRPCExecutionException
+            (.execute drpc-handler "ArbitraryDRPCFunctionName" "no-args"))))))